+++ FSNode/FSNodeRep.h
@@ -1,6 +1,6 @@
 /* FSNodeRep.h
  *  
- * Copyright (C) 2004 Free Software Foundation, Inc.
+ * Copyright (C) 2004-2013 Free Software Foundation, Inc.
  *
  * Author: Enrico Sersale <enrico@imago.ro>
  * Date: March 2004
@@ -25,8 +25,8 @@
 #ifndef FSNODE_REP_H
 #define FSNODE_REP_H
 
-#include <Foundation/Foundation.h>
-#include "FSNode.h"
+#import <Foundation/Foundation.h>
+#import "FSNode.h"
 
 typedef enum FSNInfoType {   
   FSNInfoNameType = 0,
@@ -115,9 +115,9 @@ typedef enum FSNSelectionMask {
 
 - (void)checkLocked;
 
-- (void)setGridIndex:(int)index;
+- (void)setGridIndex:(NSUInteger)index;
 
-- (int)gridIndex;
+- (NSUInteger)gridIndex;
 
 - (int)compareAccordingToName:(id)aObject;
